moving exception classes to separate module
