+++ docs/panos-upgrade-assurance/api/firewall_proxy.md
@@ -5,34 +5,6 @@ title: firewall_proxy
 hide_title: true
 custom_edit_url: null
 ---
-## class `CommandRunFailedException`
-
-Used when a command run on a device does not return the `success` status.
-
-## class `MalformedResponseException`
-
-A generic exception class used when a response does not meet the expected standards.
-
-## class `DeviceNotLicensedException`
-
-Used when no license is retrieved from a device.
-
-## class `ContentDBVersionsFormatException`
-
-Used when parsing Content DB versions fail due to an unknown version format (assuming `XXXX-YYYY`).
-
-## class `PanoramaConfigurationMissingException`
-
-Used when checking Panorama connectivity on a device that was not configured with Panorama.
-
-## class `WrongDiskSizeFormatException`
-
-Used when parsing free disk size information.
-
-## class `UpdateServerConnectivityException`
-
-Used when connection to the Update Server cannot be established.
-
 ## class `FirewallProxy`
 
 Class representing a Firewall.
